+++ crates/teloxide-core/src/bot/api.rs
@@ -1453,7 +1453,7 @@ impl Requester for Bot {
     fn gift_premium_subscription(
         &self,
         user_id: UserId,
-        month_count: u32,
+        month_count: u8,
         star_count: u32,
     ) -> Self::GiftPremiumSubscription {
         Self::GiftPremiumSubscription::new(
