+++ crates/teloxide-core/src/requests/requester.rs
@@ -1261,7 +1261,7 @@ pub trait Requester {
     fn gift_premium_subscription(
         &self,
         user_id: UserId,
-        month_count: u32,
+        month_count: u8,
         star_count: u32,
     ) -> Self::GiftPremiumSubscription;
